# Speed up the matmul op, use the gemm replace the batch gemm