matmul性能优化
1.性能对比1 
X = [1024, 10, 12]
Y = [12, 10]

未优化前
matmul gpu_time  0.591ms

优化后
matmul gpu_time  0.326ms

速度提升 1.8x

2.性能对比2 
X = [128, 50, 1000]
Y = [1000, 1000]

未优化前
matmul gpu_time  20.9ms 

优化后
matmul gpu_time  15.6ms

速度提升 1.34x

3.性能对比3
X = [512, 50, 1000]
Y = [1000, 1000]

未优化前
matmul gpu_time  61.3ms 

优化后
matmul gpu_time  43.2ms

速度提升 1.41x